+++ src/services/ExchangeRateConversionService.java
@@ -3,7 +3,7 @@ package services;
 import dto.ExchangeRateDTO;
 
 public class ExchangeRateConversionService {
-    public double convertAccordingToExchangeRate(int baseQuantity, ExchangeRateDTO exchangeRateDTO) {
+    public double convertAccordingToExchangeRate(double baseQuantity, ExchangeRateDTO exchangeRateDTO) {
         return baseQuantity * exchangeRateDTO.conversionRate();
     }
 }
